Se agregan Controller Rests, Request y Response, Services y los HTML.
- Controladores rest de Cliente, Venta e Item,
- Request y response de Cliente, Item, VentaEfectivo y Venta Tarjeta.
- Servicios de Cliente, Item y Venta.
- HTML de Clientes y Ventas (Solo lista este ultimo).
- Modificaciones en los Request y Response de Prenda.

+++ src/test/java/ar/edu/davinci/dvds20202cg2/service/PrendaServiceImplTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -27,8 +28,7 @@ class PrendaServiceImplTest {
 
 	@Autowired
 	private PrendaService prendaService;
-
-
+	
 	@Test
 	void testListAll() {
 		
@@ -62,19 +62,31 @@ class PrendaServiceImplTest {
 
 	@Test
 	void testFindById() {
-		Prenda prenda = prendaService.findById(4L);
-		
+		Optional<Prenda> prendaOptional = prendaService.findById(4L);
+		Prenda prenda = null;
+		if (prendaOptional.isPresent()) {
+			LOGGER.info("LA PRENDA FUE ENCONTRADA");
+			prenda = prendaOptional.get();
+		} else {
+			LOGGER.info("LA PRENDA NO FUE ENCONTRADA");
+		}			
 		assertNotNull(prenda);
 		assertEquals(prenda.getDescripcion(), "Pantalon Gabardina Beige");
 	}
 
 	@Test
 	void testFindById_withError() {
-		Prenda prenda = prendaService.findById(0L);
-		
+		Optional<Prenda> prendaOptional = prendaService.findById(0L);
+		Prenda prenda = null;
+		if (prendaOptional.isPresent()) {
+			LOGGER.info("LA PRENDA FUE ENCONTRADA");
+			prenda = prendaOptional.get();
+		} else {
+			LOGGER.info("LA PRENDA NO FUE ENCONTRADA");
+		}		
 		assertNull(prenda);
 	}
-	/*
+
 	@Test
 	void testSave() {
 		
@@ -92,6 +104,8 @@ class PrendaServiceImplTest {
 		assertEquals(prenda.getDescripcion(), prendaCreada.getDescripcion());
 
 		LOGGER.info("Prenda count después insert: " + prendaService.count());
+		
+		
 	}
 
 	@Test
@@ -99,11 +113,15 @@ class PrendaServiceImplTest {
 		
 		LOGGER.info("Prenda count antes delete: " + prendaService.count());
 
-		Prenda prenda = prendaService.findById(2L);
-		prendaService.delete(prenda);
-		
-		LOGGER.info("Prenda count después delete: " + prendaService.count());		
+		Optional<Prenda> prendaOptional = prendaService.findById(2L);
+		Prenda prenda = null;
+		if (prendaOptional.isPresent()) {
+			LOGGER.info("LA PRENDA FUE ENCONTRADA");
+			prenda = prendaOptional.get();
+			prendaService.delete(prenda);
+			LOGGER.info("Prenda count después delete: " + prendaService.count());
+		} else {
+			LOGGER.info("LA PRENDA NO FUE ENCONTRADA");
+		}
 	}
-	*/
 }
-
